IPA-4505: Bug fix to make code C98 compliant

+++ interop/logic/plot/plot_sample_qc.h
@@ -7,6 +7,7 @@
  */
 #pragma once
 #include <map>
+#include <math.h>
 #include "interop/util/statistics.h"
 #include "interop/constants/enums.h"
 #include "interop/logic/utils/enums.h"
@@ -112,7 +113,7 @@ namespace illumina { namespace interop { namespace logic { namespace plot {
         data.set_range(data.x_axis().min(),
                        static_cast<float>(data[0].size()+1),
                        data.y_axis().min(),
-                       std::round(max_height+5));
+                       roundf(max_height+5));
     }
